movie recommendation request insertion fixed

+++ admin/review_request_function.php
@@ -27,9 +27,19 @@
         $movie_name = $row['movie_name'];
         $mpaa = $row['mpaa'];
         $running_time = $row['running_time'];
+
+
+        $is_data_exists = $con->query("SELECT * FROM movie_summary WHERE movie_id='$id' LIMIT 1") or die($con->error());
+        $count = mysqli_num_rows($is_data_exists);
         
-        $con->query("INSERT INTO movie_summary (movie_id, movie_name, mpaa, running_time) VALUES('$id', '$movie_name', '$mpaa', '$running_time');") or die($con->error());
-        $con->query("DELETE FROM movie_recommendation_db WHERE id=$id") or die($con->error());
+        if($count == 1){
+            echo "Requesting movie data already exists in movie_summary";
+        }
+        else{
+            $insert = $con->query("INSERT INTO movie_summary (movie_id, movie_name, mpaa, running_time) VALUES('$id', '$movie_name', '$mpaa', '$running_time');") or die($con->error());
+            $con->query("DELETE FROM movie_recommendation_db WHERE id=$id") or die($con->error());
+        }
+
 
         header("location: review_request.php");
         $con->commit();
